cleanup: simple clean of "comparison between signed and unsigned errors"
not much to see here, please review but these are simple changes.



SVN revision: 52426

+++ legacy/evas/src/lib/cserve/evas_cs_client.c
@@ -515,7 +515,8 @@ evas_cserve_raw_info_get(void)
    if (!cserve) return NULL;
    if (!server_send(cserve, 0, OP_GETINFO, 0, NULL)) return NULL;
    rep = (Op_Getinfo_Reply *)server_read(cserve, 0, &opcode, &size);
-   if ((rep) && (opcode == OP_GETINFO) && (size >= sizeof(Op_Getinfo_Reply)))
+   if ((rep) && (opcode == OP_GETINFO) &&
+       (size >= (int)sizeof(Op_Getinfo_Reply)))
      {
         return rep;
      }
